+++ examples/translation/rejected/dynamic_delay_qiskit.py
@@ -1,0 +1,10 @@
+from qiskit import QuantumCircuit
+
+
+def get_duration():
+    return 8
+
+
+duration = get_duration()
+circuit = QuantumCircuit(1)
+circuit.delay(duration, 0, unit="dt")